controle-01: wip

+++ database/migrations/2024_02_26_173750_create_receipts_table.php
@@ -12,7 +12,7 @@ return new class () extends Migration {
     {
         Schema::create('receipts', function (Blueprint $table) {
             $table->id();
-            $table->float('value');
+            $table->decimal('value', 10, 2);
             $table->string('payer');
             $table->string('cpf_payer')->nullable();
             $table->string('referent');
